Fungsi POST dan Week Mode sudah,cuman Title calendar ketika Week Mode tidak jalan,NEXT:Benerin Title Calendar di Week Mode

+++ Project/app/src/main/java/com/example/CH2_PS020/fitsync/api/response/BMIResponse.kt
@@ -24,13 +24,13 @@ data class BmisItem(
 	val createdAt: String? = null,
 
 	@field:SerializedName("weight")
-	val weight: Int? = null,
+	val weight: Float? = null,
 
 	@field:SerializedName("id")
 	val id: Int? = null,
 
 	@field:SerializedName("height")
-	val height: Int? = null,
+	val height: Float? = null,
 
 	@field:SerializedName("updatedAt")
 	val updatedAt: String? = null
